+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Work around KaTeX rendering bug for `\backslash` that produced broken HTML (`&lt;/span&gt;`), triggering mdBook unclosed tag warnings.
+
 ## [0.10.0](https://github.com/lzanini/mdbook-katex/compare/v0.9.4...v0.10.0) - 2025-11-28
 
 ### Other